lambda/authorizer/requirements.txt
PyJWT==1.7.1
python-jose==3.3.0
starlette==0.37.2